refactor: extract useTestDocBoard hook and TestDocStoryPane, dedupe classification parsing, fix background-result editor flip (BRDG-426)
